rename nomodule/module dist to nomodule/module build
