+++ .travis.yml
@@ -46,7 +46,6 @@ stages:
   - name: test
   - name: experiment
     if: type = push
-  - name: bundlesize
 jobs:
   include:
     - stage: build
@@ -62,13 +61,17 @@ jobs:
       script:
         - unbuffer node build-system/pr-check/validator-tests.js
     - stage: build
-      name: 'Nomodule Dist'
+      name: 'Nomodule Build'
       script:
-        - unbuffer node build-system/pr-check/nomodule-dist.js
+        - unbuffer node build-system/pr-check/nomodule-build.js
     - stage: build
-      name: 'Module Dist'
+      name: 'Module Build'
       script:
-        - unbuffer node build-system/pr-check/module-dist.js
+        - unbuffer node build-system/pr-check/module-build.js
+    - stage: test
+      name: 'Bundle Size'
+      script:
+        - unbuffer node build-system/pr-check/bundle-size.js
     - stage: test
       name: 'Visual Diff Tests'
       script:
@@ -105,10 +108,6 @@ jobs:
       name: 'Performance Tests'
       script:
         - unbuffer node build-system/pr-check/performance-tests.js
-    - stage: bundlesize
-      name: 'Bundle Size'
-      script:
-        - unbuffer node build-system/pr-check/bundle-size.js
   allow_failures:
     - script: unbuffer node build-system/pr-check/performance-tests.js # See #28148
   fast_finish: true
