+++ build-system/pr-check/module-build.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 The AMP HTML Authors. All Rights Reserved.
+ * Copyright 2020 The AMP HTML Authors. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,31 +17,28 @@
 
 /**
  * @fileoverview
- * This script builds the minified AMP runtime.
- * This is run during the CI stage = build; job = Nomodule Dist.
+ * This script builds the esm minified AMP runtime.
+ * This is run during the CI stage = build; job = Module Dist.
  */
 
 const colors = require('ansi-colors');
 const {
   printChangeSummary,
-  processAndUploadDistOutput,
   startTimer,
   stopTimer,
   stopTimedJob,
-  timedExecWithError,
   timedExecOrDie: timedExecOrDieBase,
-  uploadDistOutput,
+  uploadEsmDistOutput,
 } = require('./utils');
 const {determineBuildTargets} = require('./build-targets');
 const {isTravisPullRequestBuild} = require('../common/travis');
 const {runNpmChecks} = require('./npm-checks');
-const {signalDistUpload} = require('../tasks/pr-deploy-bot-utils');
 
-const FILENAME = 'nomodule-dist.js';
+const FILENAME = 'module-build.js';
 const FILELOGPREFIX = colors.bold(colors.yellow(`${FILENAME}:`));
 const timedExecOrDie = (cmd) => timedExecOrDieBase(cmd, FILENAME);
 
-async function main() {
+function main() {
   const startTime = startTimer(FILENAME, FILENAME);
   if (!runNpmChecks(FILENAME)) {
     stopTimedJob(FILENAME, startTime);
@@ -50,40 +47,20 @@ async function main() {
 
   if (!isTravisPullRequestBuild()) {
     timedExecOrDie('gulp update-packages');
-    timedExecOrDie('gulp dist --fortesting');
-    uploadDistOutput(FILENAME);
+    timedExecOrDie('gulp dist --esm --fortesting');
+    uploadEsmDistOutput(FILENAME);
   } else {
     printChangeSummary(FILENAME);
     const buildTargets = determineBuildTargets(FILENAME);
-    if (
-      buildTargets.has('RUNTIME') ||
-      buildTargets.has('FLAG_CONFIG') ||
-      buildTargets.has('INTEGRATION_TEST') ||
-      buildTargets.has('E2E_TEST') ||
-      buildTargets.has('VISUAL_DIFF') ||
-      buildTargets.has('UNIT_TEST')
-    ) {
+    if (buildTargets.has('RUNTIME') || buildTargets.has('FLAG_CONFIG')) {
       timedExecOrDie('gulp update-packages');
-
-      const process = timedExecWithError('gulp dist --fortesting', FILENAME);
-      if (process.status !== 0) {
-        const error = process.error || new Error('unknown error, check logs');
-        console./*OK*/ log(colors.red('ERROR'), colors.yellow(error.message));
-        await signalDistUpload('errored');
-        stopTimedJob(FILENAME, startTime);
-        return;
-      }
-
-      timedExecOrDie('gulp storybook --build');
-      await processAndUploadDistOutput(FILENAME);
+      timedExecOrDie('gulp dist --esm --fortesting');
+      uploadEsmDistOutput(FILENAME);
     } else {
-      await signalDistUpload('skipped');
-
       console./*OK*/ log(
         `${FILELOGPREFIX} Skipping`,
-        colors.cyan('Nomodule Dist'),
-        'because this commit does not affect the runtime, flag configs,',
-        'integration tests, end-to-end tests, or visual diff tests.'
+        colors.cyan('Module Build'),
+        'because this commit does not affect the runtime or flag configs.'
       );
     }
   }
